+++ lib/shared/src/chat/recipes/helpers.ts
@@ -5,8 +5,6 @@ import { ContextMessage, getContextMessageWithResponse } from '../../codebase-co
 import { NUM_CODE_RESULTS, NUM_TEXT_RESULTS } from '../../prompt/constants'
 import { populateCodeContextTemplate } from '../../prompt/templates'
 
-import { PROMPT_TOPICS } from './fixup'
-
 export const MARKDOWN_FORMAT_PROMPT = 'Enclose code snippets with three backticks like so: ```.'
 
 const EXTENSION_TO_LANGUAGE: { [key: string]: string } = {
@@ -70,21 +68,6 @@ export async function getContextMessagesFromSelection(
 
 export function getFileExtension(fileName: string): string {
     return path.extname(fileName).slice(1).toLowerCase()
-}
-
-// This cleans up the code returned by Cody based on current behavior
-// ex. Remove  `tags:` that Cody sometimes include in the returned content
-// It also removes all spaces before a new line to keep the indentations
-export function contentSanitizer(text: string): string {
-    const FIXUP_TAG_TOPICS = `(${PROMPT_TOPICS.OUTPUT}|${PROMPT_TOPICS.SELECTED}|${PROMPT_TOPICS.PRECEDING})`
-    const FIXUP_TAG_REGEX = new RegExp(`^\\s*<${FIXUP_TAG_TOPICS}>|<\\/${FIXUP_TAG_TOPICS}>\\s*$`, 'g')
-    let output = text.replaceAll(FIXUP_TAG_REGEX, '')
-    const tagsIndex = text.indexOf('tags:')
-    if (tagsIndex !== -1) {
-        // NOTE: 6 is the length of `tags:` + 1 space
-        output = output.slice(tagsIndex + 6)
-    }
-    return output.replace(/^\s*\n/, '')
 }
 
 export const numResults = {
